refactor(06-03): extract per-user quota read into shared quota-read.ts (D-03)
- New quota-read.ts is the single home for loadUserQuota(userId) — the one
  code path /sources and /audit both consume, so the surfaces can never drift
- Move loadQuotaPlatforms / loadRedditQuota + QuotaPlatformView / RedditQuotaView
  out of sources-page-read.ts (re-exported for existing importers); zero
  behavioral change to /sources
